added restricted middleware and restructured the files
