+++ package.json
@@ -11,8 +11,11 @@
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "connect-session-knex": "^1.4.0",
     "cors": "^2.8.5",
+    "dotenv": "^8.0.0",
     "express": "^4.17.1",
+    "express-session": "^1.16.2",
     "helmet": "^3.19.0",
     "knex": "^0.19.0",
     "sqlite3": "^4.0.9"
